requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
deepmultilingualpunctuation==1.0.1
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
PyAudio==0.2.14
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.1
typing_extensions==4.13.1
urllib3==2.3.0
